small touches

+++ contracts/polygon/templates/libraries/Utils.sol
@@ -3,8 +3,8 @@ pragma solidity ^0.8.19;
 
 library Utils {
     function convertToWei(uint value)
-    internal pure
-    returns (uint256) {
+        internal pure
+        returns (uint) {
         return value * (10**18);
     }
 }
